# docs/api.openapi.yaml
openapi: 3.0.3
info:
  title: Collections Autodialer API
  version: 0.1.0
servers:
  - url: http://localhost:8000
paths:
...
  /campaigns:
    get:
      summary: List campaigns
      responses:
        '200':
          description: OK
    post:
      summary: Create campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CampaignCreate'
      responses:
        '201':
          description: Created
  /campaigns/{campaign_id}/schedule:
    post:
      summary: Schedule campaign contacts
      parameters:
        - in: path
          name: campaign_id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleRequest'
      responses:
        '202':
          description: Accepted
...
components:
  schemas:
...
    CampaignCreate:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [VOICE, EMAIL, MIXED]
        status:
          type: string
      required: [name, type]
...
    ScheduleRequest:
      type: object
      properties:
        idempotency_key:
          type: string
        correlation_id:
          type: string
      required: [idempotency_key]